% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_utils.R
\name{get_random_seed}
\alias{get_random_seed}
\alias{set_random_seed}
\alias{next_random_seed}
\alias{is_valid_random_seed}
\alias{is_lecyer_cmrg_seed}
\alias{as_lecyer_cmrg_seed}
\title{Tools for Working with Parallel Random Seeds}
\usage{
get_random_seed()

set_random_seed(seed)

next_random_seed(seed = get_random_seed())

is_valid_random_seed(seed)

is_lecyer_cmrg_seed(seed)

as_lecyer_cmrg_seed(seed)
}
\arguments{
\item{seed}{A random seed}
}
\value{
\code{get_random_seed()} returns the \emph{current} \code{.Random.seed}.  If it does not
exists, it returns \code{NULL}.

\code{set_random_seed(seed)} sets a new value on \code{.Random.seed}, and invisibly
returns the \emph{old} seed.  If \code{seed = NULL}, then the \code{.Random.seed} is
removed.

\code{next_random_seed()} updates \code{.Random.seed} by drawning an dummy random
number internally, and invisibly returns the \emph{new} seed.

\code{is_valid_random_seed(seed)} returns TRUE if \code{seed} is a valid random seed
of any RNG kind, otherwise FALSE.
This function does \emph{not} update \code{.Random.seed}.

\code{is_lecyer_cmrg_seed(seed)} returns TRUE if \code{seed} is a valid random seed
of kind \verb{L'Ecuyer-CMRG}, otherwise FALSE.
This function does \emph{not} update \code{.Random.seed}.

\code{as_lecyer_cmrg_seed(seed)} returns \verb{L'Ecuyer-CMRG} random seed based on
random seed \code{seed}.  If \code{seed} is already of the right RNG kind, then that
seed is returned as-is.  If a scalar, then a \verb{L'Ecuyer-CMRG} random seed
is generated from that seed with the help of \code{set.seed()}.
This function does \emph{not} update \code{.Random.seed}.
}
\description{
Tools for Working with Parallel Random Seeds
}
\seealso{
For more information on random number generation (RNG) in R,
see \link[base:Random]{base::Random}.
}
