% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_globals_and_packages_xapply.R
\name{get_globals_and_packages_xapply}
\alias{get_globals_and_packages_xapply}
\title{Identify Globals and Packages of a Map-Reduce Function Call}
\usage{
get_globals_and_packages_xapply(
  FUN,
  args = NULL,
  args_name = "...",
  globals = TRUE,
  packages = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{FUN}{A \link[base:function]{function} that takes one or more
arguments.}

\item{args}{(optional) A list of arguments passed to \code{FUN}, either via
a named argument (\code{args_name}), or via \dots.}

\item{args_name}{If \code{"..."}, then the arguments in \code{args} are passed
to \code{FUN()} as individual arguments.  If a string, then \code{args} as
passed to \code{FUN()} via the argument of this name.}

\item{globals}{(optional) a logical, a character vector, a named list, or a \link[globals]{Globals} object.  If TRUE, globals are identified by code inspection based on \code{expr} and \code{tweak} searching from environment \code{envir}.  If FALSE, no globals are used.  If a character vector, then globals are identified by lookup based their names \code{globals} searching from environment \code{envir}.  If a named list or a Globals object, the globals are used as is.}

\item{packages}{(optional) a character vector specifying packages
to be attached in the \R environment evaluating the future.}

\item{envir}{The \link[base:environment]{environment} from where
globals should be searched.}
}
\value{
A names list with elements:
\itemize{
\item \code{globals} - a \link[future:FutureGlobals]{FutureGlobals} object
\item \code{packages} - a character vector of package names
}
}
\description{
Identify Globals and Packages of a Map-Reduce Function Call
}
